docs: document lazy indexing (#2188)
**What type of PR is this?**

> Documentation

**What package or component does this PR mostly affect?**

> all

**What does this PR do? Why is it needed?**

In README.rst, added a subsection under `fix` and `update` explaining
how to use lazy indexing from a user's perspective. This is mentioned in
a few other places with the `-r` and `-index` flags and the `go_search`
and `proto_search` directives, but it's hard to put it all together. A
larger documentation refactoring may be needed to separate reference
material from "how to" (following [Diataxis](https://diataxis.fr/)), but
this adds the needed material for now.

In extend.md, added a section on how to implement lazy indexing in a
language extension, using Go as an example. I didn't point to code here;
not sure how useful it would be.

**Which issues(s) does this PR fix?**

For #1891

**Other notes for review**