[bzlmod/go_deps] Clarify support for modules that differ only by case. (#2213)
**What type of PR is this?**

Bug fix / minor enhancement

**What package or component does this PR mostly affect?**

bzlmod go_deps/go_repository

**What does this PR do? Why is it needed?**

Go tooling allows go.mod files to contain entries that differ only by
case.
If that happens the `go_deps` extension will `fail(..)` with an error
message describing the problem.

~In this PR we're adding support for the simple case, where the
dependency is at the same version (as is the case
[here](https://github.com/apognu/gocal/blob/master/go.mod#L6)).~

In this PR we're clarifying the error message so it doesn't suggest that
this is desired/planned functionality.

**Which issues(s) does this PR fix?**

Seems minor enough to not need an issue?

Co-authored-by: Martin Kosiba <mkosiba@apple.com>
Co-authored-by: Jay Conrod <jay@engflow.com>