walk: fix race condition (#2238)
**What type of PR is this?**

Bug fix

**What package or component does this PR mostly affect?**

walk

**What does this PR do? Why is it needed?**

Previously in Walk2, after visiting directories listed on the command
line but before visiting RelsToIndex, we set walker.mode to
UpdateDirsMode to avoid further recursion. This write raced with a read
of walker.mode in walker.shouldVisit when called by populateCache.

This change removes the walker.mode field and instead passes the mode as
an explicit parameter to functions that need it.

**Which issues(s) does this PR fix?**

Fixes #2235

**Other notes for review**

Also added a test configuration for race detection so we can catch
problems like this earlier.