feature: generate `pgoprofile` attribute if default profile detected (#2241)
<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, please read CONTRIBUTING.md and sign the
CLA
   first. We cannot review code without a signed CLA.
2. Please file an issue *first*. All features and most bug fixes should
have
an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
be updated. If you're unsure what to update, send the PR, and we'll
discuss
   in review.
-->

**What type of PR is this?**

Feature

**What package or component does this PR mostly affect?**

`language/go`

**What does this PR do? Why is it needed?**

This PR adds a quick detection of `default.pgo`, and if it exists,
automatically includes it in the `go_binary` `pgoprofile` attribute.
This is a very simple addition to allow easy generation of the
default-case for profile-guided optimization.

> _Why?_
> Golang has supported profile-guided optimizations during compilation
since `v1.20`. When using `go build`, the default value of `-pgo` is
`auto`, which will cause Go to perform profile-guided optimization if a
pprof file named `default.pgo` is co-located in the `main` package
directory. See [Go's Profile-Guided Optimization
documentation](https://go.dev/doc/pgo) for more details.

**Which issues(s) does this PR fix?**

Fixes #2240

**Other notes for review**

Hello!